+++ packages/themes/src/componentStylesOverrides/MuiTabs.ts
@@ -4,6 +4,7 @@ import * as BLUIColors from '@brightlayer-ui/colors';
 export default {
     styleOverrides: {
         root: ({ theme }) => ({
+            boxShadow: theme.vars.palette.shadows.level2,
             ...theme.applyStyles('dark', {
                 color: theme.vars.palette.text.secondary,
                 backgroundColor: BLUIColors.darkBlack[300],